(PC-6560) Dissociate tests LocationPicker - SuggestedPlaces
